Reuse the shared menu inlines in mnsound and mndatadel
mnSound_InitCenterText was a copy of Menu_InitCenterText that differed only in taking s32 rather than u8; widening the shared inline's parameter to s32 keeps every existing consumer byte-identical and lets mnsound drop the copy.

mndatadel's cooldown branch is Menu_DecrementAnimTimer written out.

mnlanguage keeps its own noinline copy, which exists to fix the .sdata2 float order.

+++ src/melee/mn/inlines.h
@@ -61,7 +61,7 @@ static inline void sfxMove(void)
     lbAudioAx_80024030(2);
 }
 
-static inline void Menu_InitCenterText(Menu* menu, u8 val)
+static inline void Menu_InitCenterText(Menu* menu, s32 val)
 {
     HSD_Text* text =
         HSD_SisLib_803A5ACC(0, 1, -9.5F, 9.1F, 17.0F, 364.68332F, 38.38772F);
